requirements.txt
streamlit
playwright==1.57.0
google-api-python-client==2.187.0
google-auth-oauthlib==1.2.3
google-auth==2.41.1
pillow
openpyxl
python-dateutil
watchdog